Anda el servo. Falta correcion ecuacion.

+++ source/MG90S.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 extern void MG90S_init(void);
-extern void MG90S_setAngle(uint8_t angle);
+extern void MG90S_setAngle(int16_t angle);
 extern void MG90S_test(void);
 
 #endif /* MG90S_H_ */
